requirements.txt
# Core deep learning frameworks
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data processing and scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Machine learning and evaluation
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.62.0
argparse

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development and testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Additional utilities
pathlib
json5
pyyaml>=5.4.0

# For model deployment (optional)
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0

# For visualization
plotly>=5.0.0
tensorboard>=2.7.0

# Memory profiling (optional)
memory-profiler>=0.58.0

# GPU monitoring (optional)
gpustat>=1.0.0
nvidia-ml-py3>=7.352.0